add legacy activity tracking
